+++ Refactor/functions/Convert-ReScriptFile.ps1
@@ -20,6 +20,10 @@
 	
 	.PARAMETER Path
 		Path to the scriptfile to modify.
+
+	.PARAMETER ProviderName
+		Name of the Token Provider to apply.
+		Defaults to: '*'
 	
 	.PARAMETER Backup
 		Whether to create a backup of the file before modifying it.
@@ -27,6 +31,12 @@
 	.PARAMETER Force
 		Whether to update files that end in ".backup.ps1"
 		By default these are skipped, as they would be the backup-files of previous conversions ... or even the current one, when providing input via pipeline!
+
+	.PARAMETER WhatIf
+		If this switch is enabled, no actions are performed but informational messages will be displayed that explain what would happen if the command were to run.
+
+	.PARAMETER Confirm
+		If this switch is enabled, you will be prompted for confirmation before executing any operations that change state.
 	
 	.EXAMPLE
 		PS C:\> Get-ChildItem C:\scripts -Recurse -Filter *.ps1 | Convert-ReScriptFile
@@ -34,13 +44,17 @@
 		Converts all scripts under C:\scripts according to the provided transformation sets.
 	#>
 	[OutputType([Refactor.TransformationResult])]
-	[CmdletBinding()]
+	[CmdletBinding(SupportsShouldProcess = $true)]
 	Param (
 		[Parameter(Mandatory = $true, ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true)]
 		[PsfValidateScript('PSFramework.Validate.FSPath.File', ErrorString = 'PSFramework.Validate.FSPath.File')]
 		[Alias('FullName')]
 		[string[]]
 		$Path,
+
+		[PsfArgumentCompleter('Refactor.TokenProvider')]
+		[string[]]
+		$ProviderName = '*',
 
 		[switch]
 		$Backup,
@@ -57,8 +71,10 @@
 			$scriptfile = [Refactor.ScriptFile]::new($file)
 			
 			try {
-				$result = $scriptfile.Transform($scriptfile.GetTokens())
-				$scriptfile.Save($Backup.ToBool())
+				$result = $scriptfile.Transform($scriptfile.GetTokens($ProviderName))
+				Invoke-PSFProtectedCommand -ActionString 'Replacing content of script' -Target $file -ScriptBlock {
+					$scriptfile.Save($Backup.ToBool())
+				} -PSCmdlet $PSCmdlet -EnableException $EnableException -Continue
 				$result
 				Write-PSFMessage -Message 'Finished processing file: {0} | Transform Count {1} | Success {2}' -StringValues $file, $result.Count, $result.Success
 			}