+++ Refactor/changelog.md
@@ -1,5 +1,14 @@
 ﻿# Changelog
 
+## ???
+
++ New: Command Get-ReToken - Scans a scriptfile for all tokens contained within.
++ Upd: Added "Line" property to all tokens
++ Upd: Get-ReTokenProvider - `-Name` parameter now accepts an array of values
++ Upd: ScriptFile - GetTokens() now supports specifying a list of providers to process
++ Upd: Convert-ReScriptFile - implements ShouldProcess
++ Upd: Convert-ReScriptFile - added `-ProviderName` parameter to support filtering by provider name
+
 ## 1.0.0 (2022-03-06)
 
 + Initial Release